Support Short, Float and BigInteger entity attribute types jmix-framework/jmix#1188

+++ jmix-cuba/cuba/src/test/groovy/spec/haulmont/cuba/core/datatypes/DatatypeTest.groovy
@@ -77,7 +77,7 @@ class DatatypeTest extends CoreTestSpecification {
 
         expect:
 
-        datatypes.find(Float.class) == null
+        datatypes.find(Byte.class) == null
     }
 
     def "adaptive BigDecimal datatype"() {
